Remove nonexistent commands
openssl help does not appear to be a command by itself. openssl help {{specific command}} is also not an option. Some commands can use the flag -help, but many of them do not support that. 

I added those spots in with some other commands for filler so it wouldn't be left with just 'openssl version'